# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WhatsApp Store Builder API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/generate-description:
    post:
      operationId: generateProductDescription
      tags: [ai]
      summary: Generate AI product description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateDescriptionBody"
      responses:
        "200":
          description: Generated description
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateDescriptionResponse"
  /ai/pricing-suggestion:
    post:
      operationId: getPricingSuggestion
      tags: [ai]
      summary: Get AI pricing suggestions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PricingSuggestionBody"
      responses:
        "200":
          description: Pricing suggestion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingSuggestionResponse"
  /ai/enhance-image:
    post:
      operationId: enhanceProductImage
      tags: [ai]
      summary: Generate an enhanced/professional product image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnhanceImageBody"
      responses:
        "200":
          description: Enhanced image as base64
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnhanceImageResponse"

  # ─── Admin ────────────────────────────────────────────────────────────────
...
  /admin/stores:
    get:
      operationId: adminListStores
      tags: [admin]
      summary: List all stores (admin only)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of stores
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStoreList"
  /admin/stores/{id}/plan:
    patch:
      operationId: adminUpdateStorePlan
      tags: [admin]
      summary: Update a store's subscription plan (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdatePlanBody"
      responses:
        "200":
          description: Plan updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
...
  /referrals/apply:
    post:
      operationId: applyReferralCode
      tags: [referrals]
      summary: Apply a referral code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyReferralBody"
      responses:
        "200":
          description: Referral applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplyReferralResponse"

  # ─── Growth Tools ─────────────────────────────────────────────────────────
...
  /growth/broadcast:
    post:
      operationId: sendBroadcast
      tags: [growth]
      summary: Generate a WhatsApp broadcast message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BroadcastBody"
      responses:
        "200":
          description: Broadcast message ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BroadcastResponse"
components:
  schemas:
...
    Subscription:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        plan:
          type: string
        status:
          type: string
        currentPeriodStart:
          type: ["string", "null"]
          format: date-time
        currentPeriodEnd:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - plan
        - status
        - createdAt
        - updatedAt
...
    GenerateDescriptionBody:
      type: object
      properties:
        productName:
          type: string
        category:
          type: ["string", "null"]
        keywords:
          type: ["string", "null"]
        tone:
          type: string
          enum: [professional, casual, luxury, fun]
      required:
        - productName
    GenerateDescriptionResponse:
      type: object
      properties:
        description:
          type: string
        shortDescription:
          type: string
      required:
        - description
        - shortDescription
    PricingSuggestionBody:
      type: object
      properties:
        productName:
          type: string
        category:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        currency:
          type: string
      required:
        - productName
        - currency
    PricingSuggestionResponse:
      type: object
      properties:
        suggestedPrice:
          type: number
        priceRange:
          type: object
          properties:
            min:
              type: number
            max:
              type: number
          required: [min, max]
        reasoning:
          type: string
      required:
        - suggestedPrice
        - priceRange
        - reasoning
    EnhanceImageBody:
      type: object
      properties:
        productName:
          type: string
        category:
          type: ["string", "null"]
        style:
          type: string
          enum: [clean_white, lifestyle, studio, minimal]
      required:
        - productName
    EnhanceImageResponse:
      type: object
      properties:
        imageUrl:
          type: string
        b64_json:
          type: string
      required:
        - imageUrl
        - b64_json

    # ─── Admin ────────────────────────────────────────────────────────────────
...
    AdminStoreItem:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        slug:
          type: string
        whatsappNumber:
          type: string
        plan:
          type: string
        productCount:
          type: integer
        orderCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - slug
        - whatsappNumber
        - plan
        - productCount
        - orderCount
        - createdAt
    AdminStoreList:
      type: object
      properties:
        stores:
          type: array
          items:
            $ref: "#/components/schemas/AdminStoreItem"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - stores
        - total
        - page
        - limit
    AdminUpdatePlanBody:
      type: object
      properties:
        plan:
          type: string
      required:
        - plan
...
    ApplyReferralBody:
      type: object
      properties:
        code:
          type: string
      required:
        - code
    ApplyReferralResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        reward:
          type: ["string", "null"]
      required:
        - success
        - message

    # ─── Growth ───────────────────────────────────────────────────────────────
...
    BroadcastBody:
      type: object
      properties:
        type:
          type: string
          enum: [promotion, new_product, announcement]
        message:
          type: string
        productIds:
          type: array
          items:
            type: integer
      required:
        - type
        - message
    BroadcastResponse:
      type: object
      properties:
        whatsappUrl:
          type: string
        message:
          type: string
        characterCount:
          type: integer
      required:
        - whatsappUrl
        - message
        - characterCount